+++ test/fixtures/case.rb
@@ -8,3 +8,8 @@ case
 when bar
   baz
 end
+%
+case # comment
+when foo
+  bar
+end